Standardise help description
Two help options (menu + icon), one is described as "About this page" and the other "Help for items on this page", which suggests they are different content - but they aren't. 

Standardise the message to be the same for both, to avoid wasting users' time.  To make sure it stays standardised, also put it in a variable so the same text is used both times if it's changed.