update system_get_sysctls() and fix bugs
update system_get_sysctls() to use $config['tunables']['item'] as an array indexed by the tunable name not arbitrary id. 

Also always return default value (for ease) and note 2 code simplifications - $value is never actually used so remove, and $disp_cache is only used for ease of finding the item given the arbitrary ID which is no longer needed, so remove that too.