Enhance pkg_mgr_install to improve calling format and allow multiple pkg actions
I'm working on an enhancement to the package manager UI, part of which will allow multiple package selection for install/remove (much like multiple rules can be selected and moved/deleted).  

At the moment a user who wants to install or remove multiple packages must action them individually, they can't select and confirm them at once (unless it's "reinstall all"). That means they can't select desired pkgs and them do something else and come back when they are all installed to view the success report, which would be nice.

However this can't be added unless pkg_mgr_install is first enhanced to allow submission of a list of packages; at the moment it only accepts a single package (or firmware), or the special action "reinstallall". As the basic functionality for multiple package action is already there ("reinstallall") adding the same for a user-specified pkg list isn't hard. 

pkg_mgr_install.php is only referenced in a very few other places in the codebase, making this easy to keep transparent.

**SCHEMA:**

I've changed the calling args for this file, to rationalise them and make it easy to provide a list of pkgs. Also because a list of pkgs would require a list of version (from, to) making _&from=XXX&to=YYY_ an impractical format anyway. This was easy as it's only called in a very few places anyway. 

Current schema in use:

	pkg_mgr_install.php?id=firmware
	pkg_mgr_install.php?mode=reinstallall
	pkg_mgr_install.php?pkg=PKGID [install pkg]
	pkg_mgr_install.php?mode=reinstallpkg&pkg=PKGID
	pkg_mgr_install.php?mode=reinstallpkg&pkg=PKGID[&from=XXX&to=YYY]
	pkg_mgr_install.php?mode=delete&pkg=PKGID

	(The "installed" mode exists but doesn't seem to be used)

It's a bit inconsistent and contains redundancy, although clear. I also think it's better the page gets the from/to versions for package updates as it does for firmware, rather than be passed them as args. 

I've changed it to just two args:

* **&action=_[firmware-upgrade | pkg-install | pkg-remove]_** - Reinstall is covered by install, and we generally say "remove" rather than "delete"

* **&pkglist=_[comma-separated list of packages to be installed/removed | all]_** - 'all' recognised as a special value here, unlikely to ever have a pkg of that name

* If needed, "installed" can be added if it serves any purpose, and no need for "from/to" as these are now obtained within the page itself