Return type+refid for the "Local Database" static authserver, that can be used compatibly with other authservers
This actually doesn't do anything at the moment, since these fields are never actually used elsewhere. But other code that handles auth servers or passes them as args (including in $_GET) might reference them or expect these values in future as they loop through the result of get_authservers_list(), or for server related logic. So let's provide some more helpful values so that calling code doesn't have to accommodate ['type'] which might have a space in it ($_GET), or 'refid' which might not be set for all servers.

Especially as I'm refactoring auth_server related code specifically to simplify the index/key by which auth servers are recognised, this is a sensible thing to be as part of that and I might rely on it later in this PR.