Formatting - indent a block of code
The next change produced a very messy diff because of indenting of a block of code. To simplify review, I've done the indent here,. with no code change whatsoever - it can be compared line number to new line number and the code is identical in this commit, except for indenting.